Extract gradle task
