+++ buildSrc/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     `kotlin-dsl`
-    `java-gradle-plugin`
 }
 
 repositories {
