+++ buildSrc/src/main/resources/META-INF/gradle-plugins/org.eclipse.dataspaceconnector.gradle.properties
@@ -1,1 +1,1 @@
-implementation-class=org.eclipse.dataspaceconnector.gradle.EDCGradlePlugin
+implementation-class=org.eclipse.dataspaceconnector.gradle.DataspaceConnectorPlugin
